Loading assets
In this change, I added some more assets (a weapon and background) to the game. I also experimented with collision between players, which failed horribly.
p5.play's methods of collision(), bounce() and displace() all add incorrectly functioning collision as one player will be able to move the other but not vice versa. This would make the game unfair and biased towards the player who can move the other one, so I decided to not include collision between players.