Update html and change game layout
I updated the HTML file to include the game's name and controls.

The game was previously a platformer with a side-on view but due to technical restrictions regarding physics and also time restrictions on the upcoming deadline, I had to change it to a top-down game. This will make it easier to develop and play, and hopefully it will allow me to complete the game when required.