Greatly increased functionality
In this commit I added weapons with shooting mechanics, health and lives, and the endgame conditions (apart from a timer) including a game over screen.

The weapon mechanics were quite easy to implement, I used a direction variable to see where the player had last moved and I use this to draw the weapon sprite on the correct side of each player. The only issue I encountered was trying to get the correct positioning, but this was solved through a bit of iterative testing.
The bullets are created from inside the barrel of the gun, making it look somewhat realistic. This system also uses the direction variable to determine where the bullet sprite is created relative to the player's position. The bullets will collide with the walls and players, damaging the players when doing so.
In terms of health and lives, the health starts at 100 and goes down by 20 per bullet hit. Each player has 5 lives, once someone loses all their lives the game transitions to the game over screen.
The game over screen is a few lines of text showing the total lives lost and who wins the game.