Added health pickup
In this commit I added the health pickup system.
Originally I intended on having the pickup respawn based on a timer, but this was too difficult to implement without breaking other components of the game. As a result, I decided to have the pickup respawn if either player dies, this was actually much easier to implement - using p5.play's sprite .removed attribute to see whether the pickup has been picked up.
As you can see, line 310 is the last remnant of my attempt at implementing the timer. If I am successful in this, then the game over screen will be able to show how much time in seconds had elapsed.